ofs: initial implementation of libofs
libofs is meant to contain primitive routines for interfacing with a dfl
feature in user space. This will typically be mmio register space.
This first implementation is more of a staging and integration into the
opae-sdk repository and build system.
This currently contains two macros that can be used to wait for
variables to either be set to a specific value or different from a
specific value.