work around missing cli11 default_val (int) function
The function
template <typename X> Option *default_val(const X &val)
Was introduced in 1.9. Previous to that, only the
Option *default_val(std::string val)
was available.

So check the version defines and convert the int to a std::string with
std::to_string().

Fixes building on rhel 8.2 where the cli11 version is 1.8

Signed-off-by: Tom Rix <trix@redhat.com>