afu_test: fix intent of default_val/default_str
CLI11::Option::default_str is documented as affecting only how the
default is presented but does not actually set the variable bounded to
the option to that value.
Most of our uses of default_val (now default_str) have the bounded
variabled defaulted but some do not. This change does one of the
following:
* Uses the bounded valriable as the input to default_str
* Uses default_val with a string input

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>