cmake: link with fmt if spdlog is found
Some distros include spdlog that uses format operations from an external
fmt library. When this is the case, finding spdlog will set
-DSPDLOG_FMT_EXTERNAL in its definitions.
This change checks for this when it finds the spdlog package that comes
as part of the distro. If it detects fmt is external, it will set
FMT_LIBRARIES to "fmt::fmt" which is the link interface used when
find_package(fmt) finds fmt on the system. When linking afu_test with
spdlog, it will check if FMT_LIBRARIES is defined and call
find_package(fmt) again from its directory so that it will import the
link target, "fmt::fmt"

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>