rpm: use cmake_* commands on fedora
The existing -B <build_dir> option in the cmake configure is a
problem on rhel.  Just removing it breaks fedora.  However when
cmake_build/install are used, it is fine.  Since the cmake_*
commands are not available on rhel 8.x, use the plain old make_*
commands.

Reduce the custom installs to a single vanilla install.
Items not installed should use the %exclude option in the files
lists.

Signed-off-by: Tom Rix <trix@redhat.com>