remove external cli11 package
cli11 is used in samples and examples, not the main library.
So having it is not a hard requirement.  Instead of fetching
a local copy, check if the header needed has already been
installed.  If it has not, do not build the effected programs

On Fedora and RHEL cli11 comes from the cli11-devel rpm.

Signed-off-by: Tom Rix <trix@redhat.com>