Master rpm (#1996)
* packaging: create-tarball.sh script

A missing file, copied from tag 2.0.0-2

Signed-off-by: Tom Rix <trix@redhat.com>

* update fedora patch number in spec file (#1865)

* update fedora patch number in spec file
  - Use an internal variable to track the release (independent of a "dist" value)
  - Update fedora patch number 2 in spec and create rpm

Signed-off-by: anandaravuri <ananda.ravuri@intel.com>

Conflicts:
	opae.spec

The main version change in same location.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: remove requires python3-sphinx

Sphinx is used when OPAE_BUILD_SPHINX_DOC option is used.
The default for this option is false and it unset in the
cmake configure step.  So the python3-sphinx is not required.
So remove.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: remove eth_group_mdev.sh from install

eth_group_mdev.sh has been removed, so it is impossible to
install.  So remove from rpm devel files.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: add n5010-ddr-test to devel files

n5010-ddr-test is built. rpmbuild fails because it is not
installed.  So install it.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: fix source url for opae tarball

Before
Source0: .../download/2.0.1-1.el8/opae-2.0.1-1.tar.gz
The '.el8' should not part of the release url, since the
release is distro independent.
Use 'opae_release' instead of 'release' in the url
After
Source0: .../download/2.0.1-1/opae-2.0.1-1.tar.gz

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: use cmake_* commands on fedora

The existing -B <build_dir> option in the cmake configure is a
problem on rhel.  Just removing it breaks fedora.  However when
cmake_build/install are used, it is fine.  Since the cmake_*
commands are not available on rhel 8.x, use the plain old make_*
commands.

Reduce the custom installs to a single vanilla install.
Items not installed should use the %exclude option in the files
lists.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: generalize check for python3-devel

On RHEL the package name is python36-devel, on fedora it is
python3-devel.  Use a glob when checking if the package is
installed.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: remove systemd-rpm-macros from prerequisite packages

systemd-rpm-macros do not exist on rhel.  On fedora, trying to remove them
produces this error.
 > rpm -e systemd-rpm-macros
error: Failed dependencies:
  systemd-rpm-macros = 246.13-1.fc33 is needed
    by (installed) systemd-246.13-1.fc33.x86_64

The existing check if systemd is installed will be enough to check that
systemd-rpm-macros are also installed on fedora.  So remove the check.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: include n5010-ddr-test.c sample code.

This sample was built, so also install its source.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: remove windows nl's in spec file

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: install missing libraries

rpmbuild on fedora complains about these libraries that are built
but not installed.  So install them.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: install new binaries

rpmbuild fails because these binaries are built but not installed.
So install them.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: add python3-jsonschema to BuildRequires

Using the create script on fedora, there is a build error

.../opae-libs/scripts/ofs/ofs_parse.py", line 33, in <module>
    import jsonschema
ModuleNotFoundError: No module named 'jsonschema'

So add python3-jsonschema to the requires list.

It was noted that python3-pip was listed twice in the create
script.  Remove one of them and reorder the list alphabetically.

Signed-off-by: Tom Rix <trix@redhat.com>

Co-authored-by: Ananda Ravuri <33236856+anandaravuri@users.noreply.github.com>
Co-authored-by: Tim Whisonant <tim.whisonant@intel.com>
Co-authored-by: Rodrigo Rojo <rodrigo.rojo@intel.com>