qpafilter: updates per request from stakeholders (#2036)
Changes the format of the input yaml file to allow for specifying
multiple integer IDs for label value and an additional adjustment
value for the upper fatal temperature threshold. The file continues
to be indexed by the string label, however the item contained at each
label is a dictionary with two keys: "id" contains a list of integer
IDs and "adjustment" contains a float that is applied to the upper
fatal threshold. Class qpamap is added to contain the in-memory data
from the yaml input.

Also adds handling for the Virtual Temperature Sensor 0, which is
constructed from the --virt-fatal-temp and --virt-warn-temp command
options.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>