Merge feature/tokens to master (#191)
Track the wrapped tokens at the opae-c level by adding them to a
list. Protect the list with a mutex for thread safety. Adds debug
logging to facilitate debug for the reference counting. Add a means
to detect leaked wrapped tokens.

Removes the properties flag OPAE_PROPERTIES_FLAG_PARENT_ALLOC and
moves instead to the new reference counting logic for wrapped
tokens.

In vfio_fpgaUpdateProperties(), don't set the parent token field
of the properties object. With the addition of ref counting for
wrapped tokens, the parent is determined at the opae-c layer so
that the search for the parent token can cross plugin boundaries.

When no parent token is available, catch the exception and
return an empty ptr_t. The caller must check the contents
of the smart pointer before using it.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>
Co-authored-by: Rodrigo Rojo <rodrigo.rojo@intel.com>