ofs_parse: fix driver init fn, field prefix (#220)
* fpagMapMMIO uses uint64_t** as its output value for mmio ptr but the
pointer offset calculations for registers within the driver struct are
caclulated in bytes. This changes the type to uint8_t* for calculating
the struct offsets but casts it to uint64_t** in the call to
fpagMapMMIO.
* The umd ast parser assumes all struct fields are prefixed with 'f_'
but the value member of the struct is not prefixed with 'f_'. This fixes
the assumption when name of field is 'value'

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>