stablizing rpm building on master (#2051)
* rpm: remove hssicommon from install

hssicommon was removed from the build with this commit
e210d53ed5b369452bb433e20243d5af385c656e

So it needs to be removed from the rpm

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: add libboard_n6010

This library is being built so install it.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: reduce globbing in the devel package

The files should be manifest of what is installed.  With globbing
the manifiest is not as clear.

Fixes the build issue of fpgadiag being listed twice.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: install ethernet bits

The hssi* scripts are installed, they need these python modules
to run.

Fixes an installed but not packaged rpm build error.

Signed-off-by: Tom Rix <trix@redhat.com>

* fpgabist: remove execution permission from source

*.cpp and *.h files are not executables, so set their
permissions for something reasonable.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: work around problem with missing build-id's

*.so's built with cmake pybind11_add_module() have
problems generating debug info because it is missing build-id's.

Signed-off-by: Tom Rix <trix@redhat.com>

* cmake: check if .git directory exists before executing git

When building from a tarball, the .git directory will not
be present.  Running git in this case produces warning noise
in the cmake configure. So if there is no .git directory, handle
as if no git executable was found.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: require cli11-devel for rpm building

To build the application that use cli11 requires that the
cli11-devel rpm be installed on Fedora and RHEL.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: require spdlog-devel for rpm building

Building the external spdlog is problematic on rhel 8.2
So use the system's spdlog-devel package instead.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: remove explict building of samples in configure

OPAE_BUILD_SAMPLES defaults to ON, so there is no need to
turn it on with the command line.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: package some python projects

rpm build was complaining that some python projects were installed
but not being packaged.  So add the projects.

Signed-off-by: Tom Rix <trix@redhat.com>

* mmlink: remove explict exit

rpmlint reports this problem
opae.x86_64: W: shared-lib-calls-exit
  /usr/lib64/libmml-srv.so.2.0.2 exit@GLIBC_2.2.5

Libraries should not exit a program.
Reviewing the usage of write_mmr, shows the access_type is always
passed in as an immediate char and is always one of the handled
values.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: exclude pybind11 build artifacts from tarball

Reveiwing the contents of the tarball shows some python
building bits are included.  a build/ and __pycache__ directory.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: exclude testing packages from the tarball

There is no testing rpm package so it is not nesseary to
include them in the tarball.

Signed-off-by: Tom Rix <trix@redhat.com>

* cmake: use file permissions when installing

Fixes an rpmlint error that gen_platform_ip.sh
was installed with 644, when it should have been 755

Signed-off-by: Tom Rix <trix@redhat.com>

* python: change shebang line from python to python3

All the py files should use python3 to avoid abiguity with
python 2

Signed-off-by: Tom Rix <trix@redhat.com>

* rpmlint: change python permission to +x

rpmlint reports an error for some *.py files.
Manually change then as recommened by the fedora project.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: add argsfilter source to install files

argsfilter.* are now being installed by cmake
So add them to the devel package.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: add systemd-devel to BuildRequires

The sample fpgaperf_counter.c includes libudev.h
Which is in systemd-devel, so require it.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: add libfpgaperf_counter.so to -devel

libfpgaperf_counter.so is built, so it needs to be
packaged.

Signed-off-by: Tom Rix <trix@redhat.com>

* rpm: add libofs_cpeng to spec file

libofs_cpeng.so is built so it needs to be installed.

Signed-off-by: Tom Rix <trix@redhat.com>

Co-authored-by: Tim Whisonant <tim.whisonant@intel.com>
Co-authored-by: Rodrigo Rojo <rodrigo.rojo@intel.com>