remove unused parameter from upcase_pci()
With clang 14.0, there is this build errror
fpgainfo.c:191:35: error: parameter 'len' set but not used
  [-Werror,-Wunused-but-set-parameter]
void upcase_pci(char *str, size_t len)
                                  ^
The len parameter is not needed.
If strcasestr can find "pci", there are at least 3 characters.
If it can not, then the null check at the top of the loop will
catch it.

Signed-off-by: Tom Rix <trix@redhat.com>