rpm: introduce the opae-*-tests.rpm
OPAE has a number of built in tests that run on either real hw
or on mock hw controlled by cmake config options.

If the user passes --define "test 1" build the tests by adding
-DOPAE_BUILD_TESTS=ON to the cmake configure step.

If the user also passes --define "mock 1" the mock tests will
be used by adding -DOPAE_BUILD_MOCK=ON

In addition to the offical unit tests, there are also misc
tests.  Move the n5010-ddr-test from -devel to -tests to start
the files list for -tests.

Signed-off-by: Tom Rix <trix@redhat.com>