PACSign: allow SR/PR keys to apply to *_TEST
Updates the key checking logic to relax the requirement on key naming
for SR/SR_TEST and PR/PR_TEST.

Previously, the strings '_fim_'/'_sr_test_' were required to appear in
the key names for SR/SR_TEST. This change relaxes the restriction so
that a key with '_fim_' can also be used for SR_TEST and vice-versa.

Likewise, the strings '_pr_'/'_pr_test_' were required to appear in
the key names for PR/PR_TEST. This change relaxes the restriction so
that a key with '_pr_' can also be used for PR_TEST and vice-versa.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>