opaevfio.py: catch write to new_id exception (#2475)
Catch any OSError exception that occurs when attempting to write
the new_id sysfs node. Print the exception and exit if it is not EEXIST,
else continue.

Catch exception generated when writing to drivers/dfl-pci/bind. If
the driver enumeration does not claim the device, this will result in
an OSError exception.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>