opaevfio, opae.io: explicitly bind to vfio-pci (#2485)
On recent kernels, eg >= 5.15, writing to new_id on the first attempt
after creating the VFs successfully binds the VF to vfio-pci. However,
after the VF is released, trying to bind it again without an explicit
write to vfio-pci's bind sysfs node fails. Handle EBUSY in the case
that the driver bind is attempted after the first write to new_id.
On subsequent writes to new_id, explicit binding is required.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>