build: reduce targets that are exported
Realizing that not all targets should be imported by external CMake
projects, I've narrowed down the list of targets to the following:
* opae-c
* opae-cxx-core
* opaemem
* ofs
* opaeuio
* opaevfio
* bitstream
* fpgaconf

For this, I've modified the opae_add_executable and
opae_add_shared_library functions to take in an optional one-value
argument to indicate the export name. I've modified the CMakeLists.txt
that are used for building the targets listing above to add the EXPORT
argument in the add target directive.

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>