cpeng: fix return code of hps command
In main for hps executable, return exit code returned by calling the
app's main function. If/when CLI11's validators determine an argument to
be invalid, the exit code returned from CLI11 will be returned.
In the case of a validation error, the error code returned is the
enumeration ValidationError which equals 105.

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>