build: export OPAEGit.cmake
* Install OPAEGit.cmake into <prefix>/lib/opae-<version>/modules
* Add the modules path to CMAKE_MODULE_PATH so that developers can
  include OPAE cmake modules

Currently only OPAEGit.cmake is being installed into this module path
(as opposed to the modules being installed in RPM packages).
Eventually, we need to only install modules that make sense for
developers. Also, we need to structure modules so that they can be
consumed by other software that is importing opae cmake configuration.
For now OPAEGit.cmake is potentially installed in two places.

Signed-off-by: Rodrigo Rojo <rodrigo.rojo@intel.com>