xfpga: sysobject: use minimum sysfs file size
An issue was identified with a sysfs file that contained a single 0.
In that case, the obj->max_size value was being set to 2, which caused
a subsequent fpgaObjectWrite() call (with size > 2) to fail. Setting
the default size large enough to accomodate most sysfs entries prevents
this issue from occurring.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>