fpgasupdate: changes related to new SDM driver
The secure device manager changed back to a sysfs-driven mechanism;
but instead of writing the firmware file name to a sysfs node, the
mechanism is to write '1' to "loading", followed by writing the binary
file to "data", followed by writing '0' to "loading". Implement this
new method alongside the existing sysfs method to preserve backwards
compatibility.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>