opae-v: one config file integration
libopae-v.so uses the same configuration table format as libopae-c.
Unlike libopae-c.so, the libopae-v.so plugin does not modify the
config data table (and it doesn't use the fields that libopae-c
modifies). With this in mind, it is ok for the two libraries to
share the same (if only the default) version of the config data
table without conflict.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>