Merge feature/config file to master (#2728)
Adds the config file parsing code for libopae-c, fpgainfo, and
fpgad to the libopae-c build.
Add "mcp" configuration to opae.cfg, as it still persists in
libopae-c's default device table.

refactor fpgad to use the one opae.cfg file API.
Remove fpgad.cfg from the installed files lists in packaging.

Print the loaded config tables in libopae-c and fpgad for debug
builds.

Refactor fpgainfo so that it acquires its platform_data_table
via the common config file parsing code.

Add a check for LIBOPAE_CFGFILE as the first possibility in
opae_find_cfg_file() in support of non-standard environments.

Implement new opae.admin.config module which loads and parses
configuration data for rsu.py and fpgareg.py at module load
time.

Implement new module opae.io.config to parse and create internal
config data for opae.io. Integrate that new module with opae.io's
ls command.

Query the value of $HOME and search in the home config file locations
relative to that path before trying the password database's version of
the home directory. This should more closely match the use of pathlib
in the Python search code.

libopae-v.so uses the same configuration table format as libopae-c.
Unlike libopae-c.so, the libopae-v.so plugin does not modify the
config data table (and it doesn't use the fields that libopae-c
modifies). With this in mind, it is ok for the two libraries to
share the same (if only the default) version of the config data
table without conflict.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>