scripts: add opae-clean.py (#2740)
Add utility script opae-clean.py to aid in scrubbing systems
when a local install from source is present. A common issue
encountered with local installs from source is that cleaning a
system of installed artifacts is nearly impossible once the
source/build tree has been removed. Using opae-clean.py,
particularly with the --deep flag can help remedy this
situation.

The script can be run from any directory, so long as the
relative positions to the project files are unchanged. For
RedHat-based installs, the script scans opae.spec.fedora in
the root of the git tree. For Debian-based installs,
packages/opae/deb/*.(install|dirs) are used.

A typical usage is to run the script without parameters, or with
the --deep parameter. The script will report what files and
directories it thinks are likely candidates to scrub; but at this
point, nothing will be removed. After you have carefully examined
the list of files and directories reported, the script can be
re-executed with the -D switch to remove the target files and
directories (given that sufficient privilege is held to do so).

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>