Fix fpgainfo events power on time stamp
SW should use the information in block with magic number '0x53696CF0' to derive power on event time stamp, and not the timestamp field of magic number '  0x53696C12'.
Power on status is logged immediately after power on.  Time of the day information is written by SW into a BMC register. This write will take some time after power on and hence with Power on log there is no timestamp value available. Because of this reason, second block 'time of the day offset' (block with magic number '0x53696CF0' ) is provided from BMC. BMC internally keeps track of time lapse after the power on and first ToD write by Host and provides the correct value by reverse calculation.  Please use this information to print the power on status time

Signed-off-by: anandaravuri <ananda.ravuri@intel.com>