Stop generating files with logs from host_exerciser (#2773)
The afu_test C++ infrastructure, used by host_exerciser, was logging both to the
console and to a separate log file. Leaving files in the working directory every
time host_exerciser is run is messy. Stop generating the files. When a log is needed,
users can pipe output to a file or use tee.