vfio: changes in support of OpenCL MMD for gzip workload (#2846)
Add a hash map implementation to libopaemem. The hash map implementation
is modeled after the implementation found in xfpga, but is a more general
implementation, operating on void pointers instead of buffer structs.

Use the hash map to replace the list-based buffer tracking in
libopaevfio to increase performance.

vfio plugin: refactor buffer info storage
Removes the linked-list tracking for the buffer info and instead stashes
the pointer to libopaevfio's underlying buffer info struct in the wsid
output parameter. This eliminates a list walk in the buffer alloc/free
paths.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>