rpm build: add unrestricted build mode
When given 'rhel' or 'fedora', the create script would check for
the presence of certain prerequisite packages before calling rpmbuild.
The RPM spec file would also require certain packages in the
'BuildRequires: ' clauses. This posed a problem for certain distros
that renamed their packages to be other than the standard package names.
eg python3-pybind11 became python39-pybind11 in one instance.

This change adds a new 'unrestricted' mode to the create script
which 1) skips the checking of prerequisite packages by the create
script and 2) removes the 'BuildRequires: ' clauses from the RPM spec.
When using 'unrestricted', it is the user's responsibility to ensure
that the SDK will build correctly so that the RPM can be created.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>