plugins: add UIO plugin (#2964)
Adds an OPAE plugin for UIO devices that is modeled after the
VFIO OPAE plugin (but uses libopaeuio.so).

* uio plugin: each token gets unique object_id
* libopaeuio: log errors only if LIBOPAE_DEBUG
Change the behavior of the ERR() macro such that it only displays
errors when LIBOPAE_DEBUG is defined. This matches the behavior
of libopaevfio.
* uio plugin: adjust fpgaReset error message
* opae-u: add unit test cases
* fpgainfo: add display for fpga_interface
* opae-u: 64-bit accesses and mmio fix
Move away from bitfield accesses to the MMIO space. These have proven
problematic in the past, eg if the compiler emits code that does a byte
access to CSR space, when the hardware requires 32-bit or 64-bit access.
Fix the BAR handling and overrun checks.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>