libopaevfio: add vf_token env var processing
When calling opae_vfio_open(), check for the presence of an environment
variable of the name "vf_token_ssss_bb_dd_f", where ssss, bb, dd, f are
the segment, bus, device, and function of the given PCIe address. When
the environment variable is found, redirect the call to
opae_vfio_secure_open(), using the value of the env var as the vf_token.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>