vfio plugin: add unit tests and correct issues (#2988)
* opae-v: add unit test cases / fix issues
* opae-v: remove bitfields, fix mmio bounds checks
Removes bitfield accesses to mmio space, replacing them with 64-bit
reads and writes. Bitfield access to mmio space has proven problematic,
eg when byte reads or writes result, but the hw requires 32-bit or
64-bit access only.
* opae-v tests: skip invalidate tests for non-mock
The tests which use the test_system instance to selectively cause
malloc/calloc to fail only work for mock targets. Skip the tests which
require this functionality when OPAE_ENABLE_MOCK is not detected.

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>