Feature: add a method for binding shared virtual addressing (#3014)
Add fpgaBindSVA() to allocate and bind PASIDs to FPGA ports. Shared
virtual addressing requires hardware support, both in the CPU and on
the FPGA.

fpgaBindSVA() depends on /dev/dfl-pci-sva from the dfl-pci driver.
IOVA-based pinned DMA remains available with older drivers and hardware
that does not support PASID.

Attempt to load dfl-pci-sva.ko in opae.io. For now, it fails silently if the
module isn't available.