Fix: correct issue pointed out by static analysis (#3044)
### Description
Static analysis complained that although a destructor had been defined in class vfio_device, there was not a corresponding copy ctor nor assignment operator. The variable is treated as a Singleton in the application, so no copying occurs.  Implementing the constructor and assignment operator as deleted solves the issue.

### Collateral (docs, reports, design examples, case IDs):


- [ ] Document Update Required? (Specify FIM/AFU/Scripts)

### Tests added:


### Tests run:
CI

Signed-off-by: Tim Whisonant <tim.whisonant@intel.com>