hssi: increate timeout to 1 second (#3132)
Some operations are taking longer than the current timeout, such as
putting an HSSI port in loopback. Timeout is a failure, so waiting
a little long isn't a problem. Actual completion observed has been
over 100ms, so chose 1 second.

The current timeout logic could be improved but is not changed in
this commit. Time passed is calculated as the sum of sleep times,
which assumes MMIO reads are free and sleep is exactly the requested
wait.

Signed-off-by: Michael Adler <michael.adler@intel.com>