opae: fix user clock configuration sequence numbers
The CSR interface for changing the user clock frequency relies on
sequence numbers to detect the next command. In normal operation,
a command stream ends by forcing the hardware sequence counter to
zero. If the command stream is interrupted, the sequence number may
be left non-zero. Update the command stream to start counting from
the current sequence number in hardware.

Signed-off-by: Michael Adler <michael.adler@altera.com>