host_exerciser: fix memory calibration check as non-root user
Commit 44f9f534afa0 ("Check memory calibration status before running
host_exerciser") introduced a pre-run check of the DDR calibration
status for each memory channel. Whereas the sysfs entries with the
calibration status are readable to non-root users, the check depends
on root privileges to open the DFL FME device. Instead of obtaining a
device handle, directly use the device token to obtain the sysfs path.

Link: https://github.com/OFS/opae-sdk/pull/3094
Link: https://hsdes.intel.com/appstore/article/#/14019997425
Link: https://hsdes.intel.com/appstore/article/#/16026664009
Signed-off-by: Peter Colberg <peter.colberg@altera.com>