userclk: add support for Agilex 5 configuration
Agilex 5 uses a new masked read-modify-write interface in hardware. OFS
defines a new general interface (type 2) that allows software to set
the address, value and mask for an IOPLL register to update. At the
moment it is used only by Agilex 5.

The M, N, C0 and C1 values for Agilex 5 and 7 are the same.

Signed-off-by: Michael Adler <michael.adler@altera.com>